requirements.txt
RPI.BME280
pytz
firebase-admin